flask/flask_init/requirements.txt
Pillow>=5.0.0
Flask>=0.12.2
Flask-WTF>=0.14.2
Flask-Mail>=0.9.1
Flask-Login>=0.4.1
Flask-Bcrypt>=0.7.1
Flask-SQLAlchemy>=2.3.2
